// sdk/src_old/domain/context/stablecoin/Capabilities.ts
export enum Capabilities {
	CASH_IN = 'Cash in',
	CASH_IN_HTS = 'Cash in hts',
	DETAILS = 'Details',
	BALANCE = 'Balance',
	BURN = 'Burn',
	BURN_HTS = 'Burn hts',
	WIPE = 'Wipe',
	WIPE_HTS = 'Wipe hts',
	FREEZE = 'Freeze',
	FREEZE_HTS = 'Freeze hts',
	PAUSE = 'Pause',
	PAUSE_HTS = 'Pause hts',
	DELETE = 'Delete',
	DELETE_HTS = 'Delete hts',
	RESCUE = 'Rescue',
	ROLE_MANAGEMENT = 'Role management',
}